users_service/requirements.txt
flask
flask-jwt-extended
werkzeug
requests
flask_sqlalchemy
psycopg2-binary
PyYAML
jsonschema